merge: 질문에 대해 AI 답변 요청 API DB 데이터 저장 로직 추가

See merge request s10-final/S10P31A506!67

+++ hellolaw/src/main/java/com/hellolaw/hellolaw/dto/QuestionAnswerResponse.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Builder
 public class QuestionAnswerResponse {
 	private String suggestion;
-	private Integer precedentId;
+	private Long precedentId;
 	private String precedentSummary;
 	private String lawType;
 	private String category;
